+++ lib/one-core/src/provider/verification_protocol/iso_mdl/verify_proof.rs
@@ -85,7 +85,7 @@ pub(crate) async fn validate_proof(
                 client_id: None,
                 response_uri: None,
                 verifier_key: None,
-                trusted_certs_skids: None,
+                trusted_certs: None,
             },
         )
         .await