+++ lib/one-core/src/service/vc_api/service.rs
@@ -321,7 +321,7 @@ impl VCAPIService {
                     response_uri: None,
                     mdoc_session_transcript: None,
                     verifier_key: None,
-                    trusted_certs_skids: None,
+                    trusted_certs: None,
                 },
             )
             .await